+++ externals/odot/libo/osc_bundle_s.h
@@ -69,8 +69,8 @@ void osc_bundle_s_setPtr(t_osc_bndl_s *bndl, char *ptr);
 t_osc_err osc_bundle_s_getMsgCount(int len, char *buf, int *count);
 t_osc_err osc_bundle_s_getMessagesWithCallback(int len, char *buf, void (*f)(t_osc_msg_s*, void *), void *context);
 
-t_osc_err osc_bundle_s_addressIsBound(int len, char *buf, char *address, int fullmatch, int *res);
-t_osc_err osc_bundle_s_addressExists(int len, char *buf, char *address, int fullmatch, int *res);
+t_osc_err osc_bundle_s_addressIsBound(long len, char *buf, char *address, int fullmatch, int *res);
+t_osc_err osc_bundle_s_addressExists(long len, char *buf, char *address, int fullmatch, int *res);
 t_osc_err osc_bundle_s_lookupAddress(int len, char *buf, char *address, t_osc_array **osc_msg_s_array, int fullmatch);
 
 t_osc_err osc_bundle_s_wrapMessage(long len, char *msg, long *bndllen, char **bndl, char *alloc);
